tests: run rust tests with std and no-std
