+++ run_tests.sh
@@ -3,6 +3,6 @@ cd golang && go clean -testcache && go test ./... -cover && \
 cd ../js && npm run test:coverage && \
 cd ../solidity && npm run test && \
 cd ../python && pipenv run test && \
-cd ../rust && cargo test && \
+cd ../rust && cargo test && carge test --no-default-features \
 cd ../c && make && \
 echo "\n\nWorks for me ¯\_(ツ)_/¯"
